Moved implementation of timing track classes to header files
