+++ RecoCTPPS/TotemRPLocal/interface/CTPPSTimingTrackRecognition.h
@@ -133,7 +133,173 @@ class CTPPSTimingTrackRecognition
     SpatialRange getHitSpatialRange(const HitVector& hits);
 };
 
-// Implementation for all member methods
-#include "RecoCTPPS/TotemRPLocal/src/CTPPSTimingTrackRecognition.cc"
+/****************************************************************************
+ * Implementation
+ ****************************************************************************/
+
+
+template<class TRACK_TYPE, class HIT_TYPE>
+CTPPSTimingTrackRecognition<TRACK_TYPE, HIT_TYPE>::CTPPSTimingTrackRecognition(const edm::ParameterSet& iConfig) :
+    threshold               ( iConfig.getParameter<double>( "threshold" ) ),
+    thresholdFromMaximum    ( iConfig.getParameter<double>( "thresholdFromMaximum" ) ),
+    resolution              ( iConfig.getParameter<double>( "resolution" ) ),
+    sigma                   ( iConfig.getParameter<double>( "sigma" ) ),
+    pixelEfficiencyFunction ( "hit_TF1_CTPPS", iConfig.getParameter<std::string>( "pixelEfficiencyFunction" ).c_str() ) {
+}
+
+
+template<class TRACK_TYPE, class HIT_TYPE>
+CTPPSTimingTrackRecognition<TRACK_TYPE, HIT_TYPE>::~CTPPSTimingTrackRecognition() {};
+
+
+template<class TRACK_TYPE, class HIT_TYPE>
+void CTPPSTimingTrackRecognition<TRACK_TYPE, HIT_TYPE>::clear() {
+  hitVectorMap.clear();
+}
+
+
+template<class TRACK_TYPE, class HIT_TYPE>
+void CTPPSTimingTrackRecognition<TRACK_TYPE, HIT_TYPE>::producePartialTracks(
+    const HitVector& hits,
+    const DimensionParameters& param,
+    float (*getHitCenter)(const HIT_TYPE&),
+    float (*getHitRangeWidth)(const HIT_TYPE&),
+    void (*setTrackCenter)(TRACK_TYPE&, float),
+    void (*setTrackSigma)(TRACK_TYPE&, float),
+    std::vector<TRACK_TYPE>& result
+  ) {
+
+  int numberOfTracks = 0;
+  const double invResolution = 1./param.resolution;
+  const float profileRangeMargin = param.sigma * 3.;
+  const float profileRangeBegin = param.rangeBegin - profileRangeMargin;
+  const float profileRangeEnd = param.rangeEnd + profileRangeMargin;
+
+  std::vector<float> hitProfile((profileRangeEnd - profileRangeBegin) * invResolution, 0.);
+  auto hitFunction = param.hitFunction;
+
+  // Creates hit profile
+  for(auto const& hit : hits) {
+
+    float center = getHitCenter(hit);
+    float rangeWidth = getHitRangeWidth(hit);
+
+    hitFunction.SetParameters(center, rangeWidth, param.sigma);
+
+    for(unsigned int i = 0; i < hitProfile.size(); ++i) {
+      hitProfile[i] += hitFunction.Eval(profileRangeBegin + i*param.resolution);
+    }
+  }
+
+  // Guard to make sure that the profile drops below the threshold at range's end
+  hitProfile.push_back(-1.);
+
+  bool underThreshold = true;
+  float rangeMaximum = -1.0;
+  bool trackRangeFound = false;
+  int trackRangeBegin = 0;
+  int trackRangeEnd;
+
+  // Searches for tracks in the hit profile
+  for(unsigned int i = 0; i < hitProfile.size(); i++) {
+
+    if(hitProfile[i] > rangeMaximum)
+      rangeMaximum = hitProfile[i];
+
+    // Going above the threshold
+    if(underThreshold && hitProfile[i] > param.threshold) {
+      underThreshold = false;
+      trackRangeBegin = i;
+      rangeMaximum = hitProfile[i];
+    }
+
+    // Going under the threshold
+    else if(!underThreshold && hitProfile[i] <= param.threshold){
+      underThreshold = true;
+      trackRangeEnd = i;
+      trackRangeFound = true;
+    }
+
+    // Finds all tracks within the track range
+    if(trackRangeFound) {
+
+      float trackThreshold = rangeMaximum - param.thresholdFromMaximum;
+      int trackBegin;
+      bool underTrackThreshold = true;
+
+      for(int j = trackRangeBegin; j <= trackRangeEnd; j++) {
+
+        if(underTrackThreshold && hitProfile[j] > trackThreshold) {
+          underTrackThreshold = false;
+          trackBegin = j;
+        }
+
+        else if(!underTrackThreshold && hitProfile[j] <= trackThreshold) {
+          underTrackThreshold = true;
+          TRACK_TYPE track;
+          float leftMargin = profileRangeBegin + param.resolution * trackBegin;
+          float rightMargin = profileRangeBegin + param.resolution * j;
+          setTrackCenter(track, (leftMargin + rightMargin) / 2.0);
+          setTrackSigma(track, (rightMargin - leftMargin) / 2.0);
+          result.push_back(track);
+          numberOfTracks++;
+        }
+      }
+
+      trackRangeFound = false;
+    }
+  }
+}
+
+
+template<class TRACK_TYPE, class HIT_TYPE>
+typename CTPPSTimingTrackRecognition<TRACK_TYPE, HIT_TYPE>::SpatialRange
+CTPPSTimingTrackRecognition<TRACK_TYPE, HIT_TYPE>::getHitSpatialRange(const HitVector& hits) {
+
+  bool initialized = false;
+  SpatialRange result;
+
+  for(const auto& hit: hits) {
+
+    if(initialized) {
+      float xBegin = hit.getX() - (hit.getXWidth() / 2.0);
+      float xEnd = hit.getX() + (hit.getXWidth() / 2.0);
+      float yBegin = hit.getY() - (hit.getYWidth() / 2.0);
+      float yEnd = hit.getY() + (hit.getYWidth() / 2.0);
+      float zBegin = hit.getZ() - (hit.getZWidth() / 2.0);
+      float zEnd = hit.getZ() + (hit.getZWidth() / 2.0);
+
+      if(xBegin < result.xBegin)
+        result.xBegin = xBegin;
+
+      if(xEnd > result.xEnd)
+        result.xEnd = xEnd;
+
+      if(yBegin < result.yBegin)
+        result.yBegin = yBegin;
+
+      if(yEnd > result.yEnd)
+        result.yEnd = yEnd;
+
+      if(zBegin < result.zBegin)
+        result.zBegin = zBegin;
+
+      if(zEnd > result.zEnd)
+        result.zEnd = zEnd;
+    }
+
+    else {
+      result.xBegin = hit.getX() - (hit.getXWidth() / 2.0);
+      result.xEnd = hit.getX() + (hit.getXWidth() / 2.0);
+      result.yBegin = hit.getY() - (hit.getYWidth() / 2.0);
+      result.yEnd = hit.getY() + (hit.getYWidth() / 2.0);
+      result.zBegin = hit.getZ() - (hit.getZWidth() / 2.0);
+      result.zEnd = hit.getZ() + (hit.getZWidth() / 2.0);
+      initialized = true;
+    }
+  }
+
+  return result;
+}
 
 #endif
